+++ base/src/main/java/io/github/sds100/keymapper/base/actions/CreateActionDelegate.kt
@@ -778,6 +778,7 @@ class CreateActionDelegate(
                         oldData.x,
                         oldData.y,
                         oldData.description ?: "",
+                        oldData.screenResolution,
                     )
                 } else {
                     null
@@ -794,6 +795,7 @@ class CreateActionDelegate(
                     result.x,
                     result.y,
                     description,
+                    result.screenResolution,
                 )
             }
 
@@ -807,6 +809,7 @@ class CreateActionDelegate(
                         oldData.fingerCount,
                         oldData.duration,
                         oldData.description ?: "",
+                        oldData.screenResolution,
                     )
                 } else {
                     null
@@ -827,6 +830,7 @@ class CreateActionDelegate(
                     result.fingerCount,
                     result.duration,
                     description,
+                    result.screenResolution,
                 )
             }
 
@@ -840,6 +844,7 @@ class CreateActionDelegate(
                         oldData.fingerCount,
                         oldData.duration,
                         oldData.description ?: "",
+                        oldData.screenResolution,
                     )
                 } else {
                     null
@@ -860,6 +865,7 @@ class CreateActionDelegate(
                     result.fingerCount,
                     result.duration,
                     description,
+                    result.screenResolution,
                 )
             }