+++ base/src/main/java/io/github/sds100/keymapper/base/actions/pinchscreen/PinchPickDisplayCoordinateFragment.kt
@@ -2,7 +2,6 @@ package io.github.sds100.keymapper.base.actions.pinchscreen
 
 import android.annotation.SuppressLint
 import android.graphics.Bitmap
-import android.graphics.Point
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -10,7 +9,6 @@ import android.view.ViewGroup
 import android.widget.ArrayAdapter
 import androidx.activity.addCallback
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.core.content.ContextCompat
 import androidx.core.os.bundleOf
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
@@ -51,12 +49,7 @@ class PinchPickDisplayCoordinateFragment : Fragment() {
 
             bitmap ?: return@registerForActivityResult
 
-            val displaySize = Point().apply {
-                @Suppress("DEPRECATION")
-                ContextCompat.getDisplayOrDefault(requireContext()).getRealSize(this)
-            }
-
-            viewModel.selectedScreenshot(bitmap, displaySize)
+            viewModel.selectedScreenshot(bitmap)
         }
 
     /**